+++ lithops/storage/backends/localhost/localhost.py
@@ -36,7 +36,7 @@ class LocalhostStorageBackend:
         logger.debug("Creating Localhost storage client")
         self.localhost_config = localhost_config
 
-        msg = STORAGE_CLI_MSG.format('Localhost')
+        msg = STORAGE_CLI_MSG.format('Localhost storage')
         logger.info("{}".format(msg))
 
     def get_client(self):